+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.1 (October 29, 2018)
+
+* Fix call to /vapid/me to use a separate api domain (e.g. api.procore.com)
+ if it is different to the auth domain (e.g. login.procore.com).
+
+  *David Hayward*
+
+
 ## 0.5.0 (October 17, 2018)
 
 * Change default site to login.procore.com (from app.procore.com)